requirements.txt
django
django-environ
mysqlclient
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
requests